Adds "cancel" button to all of the therapy settings editors
Per the slack thread, it was determined that implying that "back" would cancel the user's edited setting, using "Cancel" was determined to be more clear.  So I've added a cancel button to each editor, but it meant also wiring a cancel() function in so that the parent NavigationLink could navigate back.