LOOP-1983: Prevent crash in MockService
Stack traces seem to indicate a crash in MockService.  Either this array is being illegally accessed from multiple threads (more likely), or it is running out of memory (less likely, but still plausible given that  says he’s set the CGM update every 5 sec and it has run for hours).