LOOP-3932: Fix memory leak when deleting CGM manager (#420)
Xcode's memory graph revealed a memory leak in `WeakSynchronizedSet`.  Calling `removeValue(forKey:)` on a Swift dictionary apparently doesn't (immediately?) free up the memory associated with the entry in the set, but `compactMapValues` will. (Note the `Value` pointed to is apparently freed, as far as I can tell, but the memory allocated by the Swift Dictionary to point at it isn't.  Since this creates a new Dictionary the memory is freed.)